fixed iP2... well almost
This was a bigger one, but had to just run with a lot of broken parts until it started to come together.

Fixed GUI stuff so upon starting the kernel, things look better. Also updates correctly with shellLoad and shellRun.
Speaking of which, shellLoad continues to correctly format hex inputs and input into Memory, but now also correctly creates and adds processes to a PCB list while updating the GUI.
shellRun properly finds the correct process and executes _CPU.cycle(), which runs through the OP codes and updates the GUI every pulse. Also implemented the OP codes.